baseline/requirements.txt
bcrypt>=4.1.1,<5.0.0
cffi>=1.16.0,<2.0.0
cryptography>=42.0.0,<43.0.0
paramiko>=3.3.1,<4.0.0
PyYAML>=6.0.1,<7.0.0
